Create header on word; make event pdf same
